+++ kroxylicious-filters/kroxylicious-record-encryption/src/test/java/io/kroxylicious/filter/encryption/crypto/ParcelTest.java
@@ -23,7 +23,6 @@ import org.apache.kafka.common.record.Record;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.Mockito;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -37,6 +36,7 @@ import io.kroxylicious.testing.filter.record.RecordTestUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -73,7 +73,7 @@ class ParcelTest {
 
         buffer.flip();
 
-        BatchAwareMemoryRecordsBuilder mockBuilder = Mockito.mock(BatchAwareMemoryRecordsBuilder.class);
+        BatchAwareMemoryRecordsBuilder mockBuilder = mock(BatchAwareMemoryRecordsBuilder.class);
         parcel.readParcel(buffer, record, (v, h) -> mockBuilder.appendWithOffset(record.offset(), record.timestamp(), record.key(), v, h));
         verify(mockBuilder).appendWithOffset(record.offset(), record.timestamp(), record.key(), expectedValue, record.headers());
         assertThat(buffer.remaining()).isZero();
@@ -146,7 +146,7 @@ class ParcelTest {
     @MethodSource("exemplarStream")
     void shouldDeserializeExpectedContentsFromBytes(String name, NamedExemplar exemplar) {
         failOnVersionWithoutExemplar(exemplar);
-        Record mock = Mockito.mock(Record.class);
+        Record mock = mock(Record.class);
         when(mock.headers()).thenReturn(new org.apache.kafka.common.header.Header[0]);
         ParcelContents expected = exemplar.deserializedParcelContents;
         try {