+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/reload/AddCluster.java
@@ -108,10 +108,13 @@ final class AddCluster implements ClusterOperation {
     private Optional<ReconfigureError> bindGatewaysWithRollback(VirtualClusterModel model) {
         List<EndpointGateway> gateways = List.copyOf(model.gateways().values());
         var bindFutures = gateways.stream()
-                .map(g -> endpointRegistry.registerVirtualCluster(g).toCompletableFuture())
+                .map(g -> endpointRegistry.registerVirtualCluster(g)
+                        .thenRun(() -> g.bindPortResolver(vn -> endpointRegistry.resolvePort(vn).toCompletableFuture().join()))
+                        .toCompletableFuture())
                 .toArray(CompletableFuture[]::new);
         try {
             CompletableFuture.allOf(bindFutures).join();
+            model.logVirtualClusterSummary();
             virtualClusterRegistry.initializationSucceeded(clusterName());
             return Optional.empty();
         }