+++ kroxylicious-runtime/src/test/java/io/kroxylicious/proxy/internal/KafkaProxyInitializerTest.java
@@ -26,11 +26,11 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.InOrder;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.hamcrest.MockitoHamcrest;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import io.netty.buffer.Unpooled;
+import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelId;
@@ -54,7 +54,6 @@ import io.kroxylicious.proxy.config.ProxyProtocolMode;
 import io.kroxylicious.proxy.config.ServiceBasedPluginFactoryRegistry;
 import io.kroxylicious.proxy.config.TargetCluster;
 import io.kroxylicious.proxy.config.tls.Tls;
-import io.kroxylicious.proxy.internal.net.Endpoint;
 import io.kroxylicious.proxy.internal.net.EndpointBinding;
 import io.kroxylicious.proxy.internal.net.EndpointBindingResolver;
 import io.kroxylicious.proxy.internal.net.EndpointReconciler;
@@ -72,10 +71,12 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.inOrder;
+import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -186,7 +187,7 @@ class KafkaProxyInitializerTest {
     void shouldResolveWhenPlainChannelActivated() throws Exception {
         // Given
         final EndpointBindingResolver bindingResolver = mock(EndpointBindingResolver.class);
-        when(bindingResolver.resolve(any(Endpoint.class), isNull())).thenReturn(bindingStage);
+        when(bindingResolver.resolve(any(Channel.class), isNull())).thenReturn(bindingStage);
         kafkaProxyInitializer = createKafkaProxyInitializer(false, bindingResolver);
         when(channelPipeline.addLast(eq("plainResolver"), plainChannelResolverCaptor.capture())).thenReturn(channelPipeline);
 
@@ -197,14 +198,14 @@ class KafkaProxyInitializerTest {
         plainChannelResolverCaptor.getValue().channelActive(channelHandlerContext);
 
         // Then
-        verify(bindingResolver).resolve(any(Endpoint.class), isNull());
+        verify(bindingResolver).resolve(any(Channel.class), isNull());
     }
 
     @Test
     void shouldRemovePlainChannelInitializerOnceComplete() throws Exception {
         // Given
         final EndpointBindingResolver bindingResolver = mock(EndpointBindingResolver.class);
-        when(bindingResolver.resolve(any(Endpoint.class), isNull())).thenReturn(bindingStage);
+        when(bindingResolver.resolve(any(Channel.class), isNull())).thenReturn(bindingStage);
         kafkaProxyInitializer = createKafkaProxyInitializer(false, bindingResolver);
         when(channelPipeline.addLast(eq("plainResolver"), plainChannelResolverCaptor.capture())).thenReturn(channelPipeline);
 
@@ -237,7 +238,7 @@ class KafkaProxyInitializerTest {
         verifyFrontendHandlerAdded(orderedVerifyer);
         verifyForwardingHandlerAdded(orderedVerifyer);
         verifyErrorHandlerAdded(orderedVerifyer);
-        Mockito.verifyNoMoreInteractions(channelPipeline);
+        verifyNoMoreInteractions(channelPipeline);
     }
 
     @ParameterizedTest
@@ -260,7 +261,7 @@ class KafkaProxyInitializerTest {
         verifyFrontendHandlerAdded(orderedVerifyer);
         verifyForwardingHandlerAdded(orderedVerifyer);
         verifyErrorHandlerAdded(orderedVerifyer);
-        Mockito.verifyNoMoreInteractions(channelPipeline);
+        verifyNoMoreInteractions(channelPipeline);
     }
 
     @ParameterizedTest
@@ -283,7 +284,7 @@ class KafkaProxyInitializerTest {
         verifyFrontendHandlerAdded(orderedVerifyer);
         verifyForwardingHandlerAdded(orderedVerifyer);
         verifyErrorHandlerAdded(orderedVerifyer);
-        Mockito.verifyNoMoreInteractions(channelPipeline);
+        verifyNoMoreInteractions(channelPipeline);
     }
 
     @ParameterizedTest
@@ -412,7 +413,7 @@ class KafkaProxyInitializerTest {
                                                               ProxyProtocolMode proxyProtocolMode,
                                                               EndpointBindingResolver bindingResolver) {
         var vcc = mock(VirtualClusterRegistry.class);
-        Mockito.lenient().when(vcc.registerConnection(any(), any())).thenReturn(true);
+        lenient().when(vcc.registerConnection(any(), any())).thenReturn(true);
         return createKafkaProxyInitializer(tls, proxyProtocolMode, bindingResolver, vcc);
     }